Introduce ConfigOptions class
