+++ tests/Plugin/Config/PhpcqConfigOptionsBuilderTest.php
@@ -129,91 +129,6 @@ final class PhpcqConfigOptionsBuilderTest extends TestCase
         $instance->describeOption($mock1);
         $instance->describeOption($mock2);
 
-        $this->assertEquals([$mock1, $mock2], $instance->getOptions());
-    }
-
-    public function testValidateConfig(): void
-    {
-        $instance = new PhpcqConfigOptionsBuilder();
-        $mock1    = $this->createMock(ConfigOptionInterface::class);
-        $mock2    = $this->createMock(ConfigOptionInterface::class);
-
-        $mock1
-            ->expects($this->atLeastOnce())
-            ->method('getName')
-            ->willReturn('foo');
-
-        $mock1
-            ->expects($this->once())
-            ->method('validateValue');
-
-        $mock2
-            ->expects($this->atLeastOnce())
-            ->method('getName')
-            ->willReturn('baz');
-
-        $mock2
-            ->expects($this->once())
-            ->method('validateValue');
-
-        $instance->describeOption($mock1);
-        $instance->describeOption($mock2);
-
-        $instance->validateConfig(['foo' => 'bar', 'baz' => 1]);
-    }
-
-    public function testUnknownConfigKeys(): void
-    {
-        $instance = new PhpcqConfigOptionsBuilder();
-        $mock1    = $this->createMock(ConfigOptionInterface::class);
-        $mock2    = $this->createMock(ConfigOptionInterface::class);
-
-        $mock1
-            ->expects($this->atLeastOnce())
-            ->method('getName')
-            ->willReturn('foo');
-
-        $mock2
-            ->expects($this->atLeastOnce())
-            ->method('getName')
-            ->willReturn('baz');
-
-        $instance->describeOption($mock1);
-        $instance->describeOption($mock2);
-
-        $this->expectException(InvalidConfigException::class);
-        $instance->validateConfig(['foo' => 'bar', 'baz' => 1, 'bar' => false]);
-    }
-
-    public function testInvalidConfig(): void
-    {
-        $instance = new PhpcqConfigOptionsBuilder();
-        $mock1    = $this->createMock(ConfigOptionInterface::class);
-        $mock2    = $this->createMock(ConfigOptionInterface::class);
-
-        $mock1
-            ->expects($this->atLeastOnce())
-            ->method('getName')
-            ->willReturn('foo');
-
-        $mock1
-            ->expects($this->once())
-            ->method('validateValue');
-
-        $mock2
-            ->expects($this->atLeastOnce())
-            ->method('getName')
-            ->willReturn('baz');
-
-        $mock2
-            ->expects($this->once())
-            ->method('validateValue')
-            ->willThrowException(new InvalidConfigException());
-
-        $instance->describeOption($mock1);
-        $instance->describeOption($mock2);
-
-        $this->expectException(InvalidConfigException::class);
-        $instance->validateConfig(['foo' => 'bar', 'baz' => 1]);
+        $this->assertCount(2, $instance->getOptions());
     }
 }
